requirements.txt
Django==4.1
djangorestframework==3.13
psycopg2==2.9.*
pytest-django==4.5.2
python-dotenv==0.21.*
